#536 Add default config to event config, change file to ts
